Added some stuff

+++ commands/reply.js
@@ -8,18 +8,19 @@ let getlongstring = message.content.split(" ");
  			around: args[1]
  		})
  		.then(messages => {
+      const Discord = require('discord.js')
+      const embed = new Discord.RichEmbed()
  			const replyToMsg = messages.first();
- 			let color = (message.channel.type !== "DM") ? message.guild.member(replyToMsg.author).highestRole.color : 0x1b5fe8;
- 			message.channel.sendMessage(`${replyToMsg.author}, ${getlongstring}`, {
- 				embed: {
- 					color: color,
- 					author: {
- 						name: `${replyToMsg.author.username} (${replyToMsg.author.id})`,
- 						icon_url: replyToMsg.author.avatarURL
- 					},
- 					description: replyToMsg.content
- 				}
- 			}).then(() => message.delete()).catch(console.error);
-
+      let attachment = '';
+      if(!!replyToMsg.attachments.first() && !!replyToMsg.attachments.first().width) {
+        attachment = replyToMsg.attachments.first().url
+      }
+    console.log(attachment)
+ 			let color = (message.channel.type !== "DM" && message.guild && message.guild.member(replyToMsg.author)) ? message.guild.member(replyToMsg.author).highestRole.color : 0x1b5fe8;
+ 			embed.setColor(color);
+      embed.setAuthor(`${replyToMsg.author.username} (${replyToMsg.author.id})`, replyToMsg.author.avatarURL);
+      embed.setDescription(replyToMsg.content);
+      embed.setImage(attachment)
+    message.channel.sendEmbed(embed,`${replyToMsg.author}, ${getlongstring}`).then(() => message.delete()).catch(console.error); 
  		}).catch(console.error);
     }}
